Literally hammer a railroad spike up my rectum. We have wiped the repo and restarted
Hi students,
There were some inconsistencies between the rubric and the project files. Some "to be implemented" tags were missing and some where in the wrong place and I had to rename a method. Also, in the movieDB file there was an input which was supposed to be a float but was in the double format. I apologize for the confusions.
- I have updated the MovieDB.java, Table.java. Please re-download them.
- I have also updated the rubric accordingly.
- In the expected output file, there are extra outputs. Please ignore them. Only match the tests that are in movieDB file.
We have extended project 1 and the peer-reviewing deadline for two days. Feel free to stop by my office hours this Friday or Monday if you need assistance or send me an e-mail.
- Rezwan

Dear Future Employer,
Please don't do this to me
Thanks in advance <3